InviteSignUpJWT no longer function component declared within App.tsx, but outside
