+++ public/index.html
@@ -34,7 +34,8 @@
     <!-- Content Security Policy -->
     <meta http-equiv="Content-Security-Policy" content="
       default-src 'self' 'unsafe-inline';
-      script-src 'self' https://ajax.cloudflare.com/
+      script-src 'self' 'unsafe-inline' 
+      https://ajax.cloudflare.com/
       http://localhost:*
       http://127.0.0.1/
       http://localhost/
