+++ src/App.tsx
@@ -8,7 +8,6 @@ import {
 } from 'react-router-dom';
 import './static/styles/App.scss';
 import { Helmet } from 'react-helmet';
-import { useParams } from 'react-router';
 import PersonSignup from './components/SignUp/PersonSignup';
 import Header from './components/Header';
 import UploadDocs from './components/UploadDocs';
@@ -43,6 +42,7 @@ import SignupBrancher from './components/SignUp/SignupBrancher';
 import Careers from './components/AboutUs/Careers';
 import InviteSignupFlow from './components/SignUp/InviteSignupFlow';
 import AdminDashboard from './components/AdminDashboard';
+import InviteSignupJWT from './components/SignUp/InviteSignupJWT';
 
 interface State {
   role: Role,
@@ -55,22 +55,7 @@ interface State {
 
 const timeUntilWarn: number = 1000 * 60 * 120;
 const timeFromWarnToLogout: number = 1000 * 60;
-const jwtDecode = require('jwt-decode');
 // const timeoutTotal: number = timeUntilWarn + timeFromWarnToLogout;
-
-function InviteSignupJWT() {
-  const { jwt } = useParams();
-  try {
-    const decoded = jwtDecode(jwt);
-    const currentTime = Date.now() / 1000;
-    if (decoded.exp > currentTime) {
-      return <InviteSignupFlow orgName={decoded.organization} role={decoded.role} />;
-    }
-  } catch (err) {
-    return <Redirect to="/error" />;
-  }
-  return <Redirect to="/error" />;
-}
 
 class App extends React.Component<{}, State, {}> {
   private idleTimerWarn;
