+++ src/components/SignUp/InviteSignupJWT.tsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import { useParams } from 'react-router';
+import {
+  BrowserRouter as Router,
+  Route,
+  Redirect,
+  Switch,
+} from 'react-router-dom';
+import InviteSignupFlow from './InviteSignupFlow';
+
+const jwtDecode = require('jwt-decode');
+
+function InviteSignupJWT() {
+  const { jwt } = useParams();
+  try {
+    const decoded = jwtDecode(jwt);
+    const currentTime = Date.now() / 1000;
+    if (decoded.exp > currentTime) {
+      return <InviteSignupFlow orgName={decoded.organization} role={decoded.role} />;
+    }
+  } catch (err) {
+    return <Redirect to="/error" />;
+  }
+  return <Redirect to="/error" />;
+}
+
+export default InviteSignupJWT;
